Partial work towards querying branch point.
Previously, when calculating the impact on HEAD builds,
ClusterFuzz would use the latest available build. For Linux builds,
where most of our test cases run, there might be a long delay
between branch point and the availability of a build.

This change is work-in-progress to instead look up the
branch point directly. It's incomplete.

Change-Id: I8a6ae5dfc29a8c6cceb5cf3a46b9fa0c9c904f40