Remove CF knowledge of branch names.
This is a first partial attempt at removing ClusterFuzz's awareness of
flawed concepts such as 'head', 'beta', 'stable', 'extended stable' etc.

These are problematic concepts for ClusterFuzz to remember because they
don't map 1:1 with actual Chrome branches: for instance,
* The same branch may be both Stable and Extended Stable
* There might not be a beta branch sometimes
* A branch might not yet be Dev or Canary but may still be important

In the rest of our Chromium security processes we've moved away from
depending on these names (via the Security_Impact label) and switched to
depending entirely on numeric branch numbers (the FoundIn label). This
does the same in ClusterFuzz.

This change is partial in the following ways which will need addressing
before merge:
* It's pseudo-code and none of this has been run
* It doesn't fix up the tests
* It doesn't attempt any form of backwards compatibility with older
  test cases. (Possibly this can be fixed by re-running the impact
  task on test cases we care about; otherwise we can tweak the boolean
  field which records whether impact has been deduced).
* It still asks Chromiumdash about relevant branches by asking about
  branch names. Ideally it would use an alternative query where all
  current relevant release branches are enumerated.
* Similarly, it doesn't attempt to identify branches which are
  in-between the named branches.
* It probably doesn't consider HEAD impact properly.